Try to fix the Windows package.
On Windows we don't make shared libraries so we were deleting the key library
file at the bottom of build.sh! Also we weren't treating libX11-xcb on the
same footing as libX11. And we were distributing a bunch of manpage that I
suspect nobody wants. Let's try to tidy that all up.