recipe/windows-configure.patch: attempt to work around autoconf version minimum
This is the same issue as seen in
https://github.com/conda-forge/xorg-libxt-feedstock/pull/13 -- it looks
like the X.org packages are bumping their autoconf requirements to 2.70,
while our Windows toolchain only provides 2.69. Unfortunately, we don't
really know how to update the Windows tools right now, as tracked in
https://github.com/conda-forge/conda-forge.github.io/issues/1891 .

Let's see if it works to just patch the version requirement down. The
2.70 release of Autoconf is billed as introducing several important
changes (and it comes 8 years after 2.69), but I would expect that if it
doesn't reject the `configure.ac` specification, we should be OK.